backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
requests==2.32.3
numpy==2.2.6
scikit-learn==1.5.2
joblib==1.4.2
python-dotenv==1.2.1
shap==0.50.0
pytz==2024.1